refactor: split PaneTreeView.swift into focused single-responsibility files
Break up the 716-line god-file into 5 focused files:
- PaneTreeView.swift — PaneTreeView + PaneSplitView (recursive tree rendering)
- PaneDividerView.swift — draggable divider between panes
- PaneLeafView.swift — single leaf pane with editor, git, and tab management
- PaneDropZone.swift — drop zone enum, overlay, preference key, drop delegate
- PaneFocusDetector.swift — NSViewRepresentable focus detection via event monitor

Private types promoted to internal where needed for cross-file access.
